Introduce Event object for late success or failure
